Feat: finalizado cadastro de despesas.

+++ Alugueis_API/Models/DespesaRateioModel.cs
@@ -4,7 +4,7 @@
     {
         public int CodDespesa { get; set; }
         public int CodApto { get; set; }
-        public float VlrRateio { get; set; }
+        public double VlrRateio { get; set; }
         public Despesa Despesa { get; set; }
         public Apto Apto { get; set; }
     }
